+++ lib/Component/VCard.php
@@ -227,17 +227,37 @@ class VCard extends VObject\Document {
                     $this->VERSION = $versionMap[self::DEFAULT_VERSION];
                 }
             }
+            if ($version === '2.1' && ($options & self::PROFILE_CARDDAV)) {
+                $warnings[] = array(
+                    'level' => 3,
+                    'message' => 'CardDAV servers are not allowed to accept vCard 2.1.',
+                    'node' => $this,
+                );
+            }
 
         }
         $uid = $this->select('UID');
-        if (($options & self::REPAIR) && count($uid) === 0) {
-            $this->UID = VObject\UUIDUtil::getUUID();
+        if (count($uid) === 0) {
+            if ($options & self::PROFILE_CARDDAV) {
+                // Required for CardDAV
+                $warningLevel = 3;
+                $message = 'vCards on CardDAV servers MUST have a UID property.';
+            } else {
+                // Not required for regular vcards
+                $warningLevel = 2;
+                $message = 'Adding a UID to a vCard property is recommended.';
+            }
+            if ($options & self::REPAIR) {
+                $this->UID = VObject\UUIDUtil::getUUID();
+                $warningLevel = 1;
+            }
             $warnings[] = array(
-                'level' => 1,
-                'message' => 'The UID property must appear in the VCARD component exactly 1 time',
+                'level' => $warningLevel,
+                'message' => $message,
                 'node' => $this,
             );
         }
+
         $fn = $this->select('FN');
         if (count($fn)!==1) {
 
@@ -330,17 +350,7 @@ class VCard extends VObject\Document {
             // FN is commented out, because it's already handled by the
             // validate function, which may also try to repair it.
             // 'FN'           => '+',
-
-            // vcard actually specifies this as '?', but in most cases not
-            // having a UID is highly undesirable. So here we're going against
-            // the spec and make it required.
-            //
-            // I would be interested to hear if this is problematic for
-            // anyone, or at least a usecase where this is undesirable.
-            //
-            // If so, I may have to add a facility that allows us to check
-            // specifically for validity in the context of 'DAV'.
-            'UID'          => '1',
+            'UID'          => '?',
         ];
 
     }