+++ ChangeLog.md
@@ -18,6 +18,13 @@ ChangeLog
 
 * #168: Expanding calendars now removes objects with recurrence rules that
   don't have a valid recurrence instance.
+* #177: SCHEDULE-STATUS should not contain a reason phrase, only a status
+  code.
+* #175: Parser can now read and skip the UTF-8 BOM.
+* #179: Added `isFloating` to `DATE-TIME` properties.
+* #179: Fixed jCal serialization of floating `DATE-TIME` properties.
+* #173: vCard converter failed for `X-ABDATE` properties that had no
+  `X-ABLABEL`.
 
 
 3.3.4 (2014-11-19)